fix: changed file at handdleProductSelect at drawer puchase
